fix(cli): extract query --storey resolution into query-storey.ts (module-size ratchet)
CI failed the module-size ratchet: query.ts grew to 623 lines against its
609-line budget after the same-named-storey union fix. Extracted the
--storey resolution (expressId/exact-name/substring matching and the
storeyIds union) into a new sibling module, packages/cli/src/commands/
query-storey.ts, exporting resolveStoreyIds(bim, storeyFilter): Set<number>.
Behavior is unchanged; query.ts now calls the extracted function and drops
to 600 lines (9 lines of headroom under the 609 budget).
