+++ packages/cli/src/commands/query.ts
@@ -13,6 +13,7 @@
 import { createHeadlessContext } from '../loader.js';
 import { printJson, getFlag, hasFlag, fatal, validateLimit } from '../output.js';
 import { STANDARD_QTO_MAP, sortEntities } from './query-aggregation.js';
+import { resolveStoreyIds } from './query-storey.js';
 import { VALID_GROUP_BY_KEYS, outputCount, outputSum, outputAggregation, outputGroupBy, outputEntities, computeUniqueValues } from './query-output.js';
 import { applyWhereFilter, parseWhereFilter, compareValues, normalizeBooleanValue } from './where-filter.js';
 
@@ -297,34 +298,10 @@ export async function queryCommand(args: string[]): Promise<void> {
     q = q.byType(...types);
   }
 
-  // --storey filter: restrict to entities in a specific storey (or storeys:
-  // IfcBuildingStorey.Name is not unique — two storeys sharing a Name are
-  // legal siblings under different buildings, or duplicate levels in a
-  // malformed/federated file. Matching by GlobalId/expressId is unambiguous;
-  // matching by Name is not, so every storey with that Name is included
-  // rather than picking an arbitrary first match and silently dropping the
-  // rest (a same-named-storey element would otherwise vanish from the
-  // result depending on array order).
+  // --storey filter: restrict to entities in a specific storey (or storeys
+  // sharing a Name — see resolveStoreyIds).
   if (storeyFilter) {
-    const storeys = bim.storeys();
-    const byExpressId = storeys.find((s: any) => String(s.ref.expressId) === storeyFilter);
-    let matchedStoreys: any[];
-    if (byExpressId) {
-      matchedStoreys = [byExpressId];
-    } else {
-      const exactNameMatches = storeys.filter((s: any) => s.name === storeyFilter);
-      matchedStoreys = exactNameMatches.length > 0
-        ? exactNameMatches
-        : storeys.filter((s: any) => s.name.toLowerCase().includes(storeyFilter.toLowerCase()));
-    }
-    if (matchedStoreys.length === 0) {
-      const names = storeys.map((s: any) => s.name).filter(Boolean).join(', ');
-      fatal(`Storey "${storeyFilter}" not found. Available: ${names || '(none)'}`);
-    }
-    const storeyIds = new Set<number>();
-    for (const storey of matchedStoreys) {
-      for (const e of bim.contains(storey.ref)) storeyIds.add(e.ref.expressId);
-    }
+    const storeyIds = resolveStoreyIds(bim, storeyFilter);
     // Post-filter: only keep entities that are in this storey
     const baseEntities = q.toArray();
     let storeyEntities = baseEntities.filter((e: any) => storeyIds.has(e.ref.expressId));
